requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
beautifulsoup4==4.14.2
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
cloudpathlib==0.22.0
cohere==5.18.0
confection==0.1.5
cymem==2.0.11
dataclasses-json==0.6.7
dill==0.4.0
docling==2.55.0
docling-core==2.48.3
docling-ibm-models==3.9.1
docling-parse==4.5.0
easyocr==1.7.2
es_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.8.0/es_core_news_sm-3.8.0-py3-none-any.whl#sha256=e451a83d6df79b87e9eed0cb553f03e99e36a3bab18a7b79f0dcfd1fdf875e12
et_xmlfile==2.0.0
Faker==37.8.0
fastavro==1.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.35.3
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langchain==0.3.27
langchain-cohere==0.4.6
langchain-community==0.3.30
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.4.31
language_data==1.3.0
latex2mathml==3.78.1
lazy_loader==0.4
lxml==5.4.0
marisa-trie==1.3.1
markdown-it-py==4.0.0
marko==2.2.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.18
murmurhash==1.0.13
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
ollama==0.6.0
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==2.22.2
preshed==3.0.10
propcache==0.3.2
psutil==7.1.0
pyclipper==1.3.0.post6
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
pypdf==6.1.1
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
rpds-py==0.27.1
rtree==1.4.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.2
semchunk==2.2.2
sentence-transformers==5.1.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tifffile==2025.9.30
tokenizers==0.22.1
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.2
typer==0.19.2
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.3
xlsxwriter==3.2.9
yarl==1.20.1
zstandard==0.25.0
pymilvus
openai